requirements.txt
# Core ML
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0

# Data handling
pandas>=2.0.0
pyarrow>=14.0.0
pyyaml>=6.0

# Evaluation & utilities
scikit-learn>=1.3.0
numpy>=1.24.0

# Serialization
joblib>=1.3.0
